Refactoring of fonts. Refactoring of mouse input logic for gui. Refactoring of engine

+++ game-logic/common/src/main/java/net/bfsr/config/component/shield/ShieldData.java
@@ -12,7 +12,7 @@ public class ShieldData extends ConfigData {
     private final Path texturePath;
     private final float maxShield;
     private final float regenAmount;
-    private final float rebuildTimeInTicks;
+    private final int rebuildTimeInTicks;
 
     ShieldData(ShieldConfig config, String fileName, int id, int registryId) {
         super(fileName, id, registryId);
